Improved Early File Closing To Avoid Race Condition
a LIMIT clause will close the data file once the limit has been
reach, to avoid reading the rest of the file unnecessarily. This
was causing an error if the file had already been closed by
reaching the end of the file normally. To avoid this, it now
keeps track of whether the read stream is still active and only
attempts to close the file if it is.